requirements.txt
django == 1.8.2
django-extensions == 1.5.5
djangorestframework == 3.1.3
envdir == 0.7

mysql-python == 1.2.5
django-debug-toolbar == 0.9.1
django-taggit==0.16.3
django-js-reverse==0.6.1


